requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
_tflow_select=2.3.0=mkl
abseil-cpp=20211102.0=hd4dd3e8_0
absl-py=1.3.0=py37h06a4308_0
aiohttp=3.8.1=py37h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
anaconda-client=1.11.0=py37h06a4308_0
anaconda-navigator=2.3.2=py37h06a4308_0
anyio=3.5.0=py37h06a4308_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py37h7f8727e_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=py37h06a4308_0
asynctest=0.13.0=py_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py37h06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py37h06a4308_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py37h27cfd23_1003
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.10.11=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.9.24=py37h06a4308_0
cffi=1.15.1=py37h74dc2b5_0
chardet=4.0.0=py37h06a4308_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py37h06a4308_0
cloudpickle=2.2.0=pypi_0
clyent=1.2.2=py37_1
conda=22.9.0=py37h06a4308_0
conda-content-trust=0.1.3=py37h06a4308_0
conda-package-handling=1.9.0=py37h5eee18b_1
conda-repo-cli=1.0.24=py37h06a4308_0
conda-token=0.4.0=pyhd3eb1b0_0
cryptography=38.0.1=py37h9ce1e76_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
dbus=1.13.18=hb2f20db_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dill=0.3.5.1=pypi_0
dm-tree=0.1.7=pypi_0
entrypoints=0.4=py37h06a4308_0
etils=0.8.0=pypi_0
expat=2.4.9=h6a678d5_0
fftw=3.3.9=h27cfd23_1
flatbuffers=22.10.26=pypi_0
fontconfig=2.13.1=hef1e5e3_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.2.0=py37h7f8727e_0
gast=0.4.0=pypi_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
googleapis-common-protos=1.56.4=pypi_0
grpc-cpp=1.46.1=h33aed49_0
grpcio=1.42.0=py37hce63b2e_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=3.7.0=py37h737f45e_0
hdf5=1.10.6=h3ffc7dd_1
icu=58.2=he6710b0_3
idna=3.4=py37h06a4308_0
imageio=2.22.4=pypi_0
importlib-metadata=4.11.3=py37h06a4308_0
importlib_metadata=4.11.3=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.2=py37h06a4308_0
ipython=7.31.1=py37h06a4308_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py37h06a4308_1
jinja2=3.1.2=py37h06a4308_0
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.16.0=py37h06a4308_0
jupyter=1.0.0=py37h06a4308_8
jupyter_client=7.3.5=py37h06a4308_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.11.2=py37h06a4308_0
jupyter_server=1.18.1=py37h06a4308_0
jupyterlab=3.4.4=py37h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.15.2=py37h06a4308_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.2=py37h295c915_0
krb5=1.19.2=hac12032_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=14.0.6=pypi_0
libcurl=7.85.0=h91b91d3_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libllvm10=10.0.1=hbcb73fb_5
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.1=h4ff587b_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_1
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py37h06a4308_0
markupsafe=2.1.1=py37h7f8727e_0
matplotlib=3.5.2=py37h06a4308_0
matplotlib-base=3.5.2=py37hf590b9c_0
matplotlib-inline=0.1.6=py37h06a4308_0
mistune=0.8.4=py37h14c3975_1001
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
multidict=6.0.2=py37h5eee18b_0
munkres=1.1.4=py_0
navigator-updater=0.3.0=py37h06a4308_0
nbclassic=0.4.8=py37h06a4308_0
nbclient=0.5.13=py37h06a4308_0
nbconvert=6.4.4=py37h06a4308_0
nbformat=5.5.0=py37h06a4308_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py37h06a4308_0
notebook=6.5.2=py37h06a4308_0
notebook-shim=0.2.2=py37h06a4308_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
oauthlib=3.2.1=py37h06a4308_0
openssl=1.1.1s=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=pypi_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathlib=1.0.1=py37_2
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py37hace64e9_1
pip=22.2.2=py37h06a4308_0
pkgutil-resolve-name=1.3.10=py37h06a4308_0
ply=3.11=py37_0
prometheus_client=0.14.1=py37h06a4308_0
promise=2.3=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.19.6=pypi_0
psutil=5.9.0=py37h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pybind11=2.10.0=pypi_0
pycosat=0.6.4=py37h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.4.0=py37h06a4308_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py37h06a4308_0
pyqt=5.15.7=py37h6a678d5_1
pyqt5-sip=12.11.0=py37h6a678d5_1
pyrsistent=0.18.0=py37heee7806_0
pysocks=1.7.1=py37_1
pystemmer=2.0.1=pypi_0
python=3.7.15=haa1d7c7_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py37h06a4308_0
python-mnist=0.7=pypi_0
pytz=2022.1=py37h06a4308_0
pyyaml=6.0=py37h7f8727e_1
pyzmq=23.2.0=py37h6a678d5_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtconsole=5.3.2=py37h06a4308_0
qtpy=2.2.0=py37h06a4308_0
qtwebkit=5.212=h4eab89a_4
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
requests=2.28.1=py37h06a4308_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
ruamel_yaml=0.15.100=py37h27cfd23_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=py37h6c91a56_2
seaborn=0.12.1=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=65.5.0=py37h06a4308_0
sip=6.6.2=py37h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
snappy=1.1.9=h295c915_0
sniffio=1.2.0=py37h06a4308_1
soupsieve=2.3.2.post1=py37h06a4308_0
sqlite=3.39.3=h5082296_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.0=py37hca6d32c_0
tensorboard-plugin-wit=1.8.1=py37h06a4308_0
tensorflow=2.10.0=pypi_0
tensorflow-datasets=4.7.0=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-gan=2.1.0=pypi_0
tensorflow-hub=0.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.27.0=pypi_0
tensorflow-metadata=1.10.0=pypi_0
tensorflow-probability=0.18.0=pypi_0
termcolor=2.1.0=py37h06a4308_0
terminado=0.13.1=py37h06a4308_0
testpath=0.6.0=py37h06a4308_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=py37h06a4308_0
tornado=6.2=py37h5eee18b_0
tqdm=4.64.1=py37h06a4308_0
traitlets=5.1.1=pyhd3eb1b0_0
typing-extensions=4.3.0=py37h06a4308_0
typing_extensions=4.3.0=py37h06a4308_0
ujson=5.4.0=py37h6a678d5_0
urllib3=1.26.12=py37h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py37_1
websocket-client=0.58.0=py37h06a4308_4
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py37h06a4308_0
wrapt=1.14.1=py37h5eee18b_0
xz=5.2.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py37h5eee18b_0
zeromq=4.3.4=h2531618_0
zipp=3.8.0=py37h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
